use a bucket array for objects
This is the first step in an effort to improve the robustness of objects
and strings in muon.  Prior to this commit, these resources were each
stored in a single contiguous array.  This made the implementation
extremely simple, but introduced subtle bugs.  When the array had to
grow, it's backing memory could potentially be moved, which would
invalidate all pointers to elements in the array.  A classic realloc()
gotcha.  This effectively meant you could not take a pointer to an
object, unless you were absolutely sure no one else would create a new
object between the time you took the pointer and when you used it.  To
make matters worse, if objects were created, but they didn't trigger a
realloc (there was still space in the array), there would be no error.

This commit creates a new data structure "bucket_array", which grows
without invalidating old memory, at the expense of fragmentation and
one more indirection.  This eliminates the above issue for objects, and
opens up the possibility of only referring to objects by pointer, and
doing away with object ids entirely.