libpkgconf: bypass list iteration when pulling variables/dependencies
List iteration is only needed for dependency graphs where multiple dependencies
are specified in the queue.  In the case of muon, the dependency resolution queue
only has a single dependency in the resolution queue in any case.

Because of various reasons (people doing crazy things in their autoconf scripts),
pkgconf has to support --modversion et al for multiple dependencies, like its
freedesktop.org counterpart, but this thankfully does not apply to muon.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>