add helpful error message when -Dopt=val is used after the builddir
This violates the POSIX Utility Syntax Guidelines and muon does not
support it. But meson *does* support GNU-style argument permutation, so
people (read: me) may forget themselves and try it anyway.

When that happens, the current behavior is to simply drop the extra
operands on the floor, silently. Elevate this to an error, and throw in
a helpful bit of advice.