don't default to auto if no subcommand is provided
The original intent of auto was to be a replacement for the `meson
build` invocation which is many people's first attempt at building a
project with meson.  Accordingly, auto was originally named build so you
could do `muon build`.  This only works with a `.muon` file currently,
but a later feature was going to be to assume all default options, and a
build directory called build when no `.muon` file is provided.

In order to make `muon auto` even _more_ ergonomic than `meson build`,
muon defaulted to `auto` if no arguments were passed.  In theory you
could just run `muon` in a directory and the project would be setup and
built with ninja with no further intervention.

Unfortunately it is a pretty bad user experience for new muon users
coming from meson to just type `muon` and get `error .muon is not a
file`.  Additionally, `.muon` files are not documented anywhere.  For
these reasons I am removing the implicit auto behavior.