run_cmd: start copying output before child exits
In order to capture the output of the child process, two pipes are
created, one for stdout and one for stderr.  Originally, the pipe was
only copied to a buffer in the parent after the child exited.  This
causes a deadlock if the child outputs more bytes than the pipe can hold
because the child will wait until the pipe can be written to, while the
parent won't read from the pipe until the child exits.

On linux, the default pipe capacity is 65,536 bytes which is rather
large, and helped keep this issue undiscovered.  On netbsd, the pipe
size appears to be only 16,384 bytes, and is where I first noticed this.