fix obj_printf va_arg handling
There was a big oversight in how varargs were handled.  All %o
specifiers were removed from the original format string, but the
original va_list was not modified, which meant that any normal (non-%o)
specifiers passed after an %o would be pointing to the wrong argument.

E.g. obj_printf("%s %o %s", "hello", (muon obj)123, "goodbye") would be
translated to something like: printf("%s [1, 2, 3] %s", "hello", (muon
obj)123, "goodbye"), and so the second %s would try to interpret 123 as
a char pointer.

The current solution uses a union to store a "generic" argument that is
then passed to a seperate printf() invocation for each encountered
specifier.  This has the advantage that the output buffer can be written
to directly, there is no need for an intermediate format buffer.