optimize plusassign for arrays
When you do array addition, a whole new array has to be created.  For
instance, [1] + [2] creates a brand new array [1, 2].  This means that
in a loop like:

foreach x : list
	y = y + x
endforeach

Every time y + x is computed, the entire array of y is copied, and x is
appended to the new array.  This is necessary for the language's
semantics to work, but it gets slower with every append.

If you use += though, there isn't a need to compute an intermediate
object and you can directly mutate the lhs.  This commit implements that
insight.