during builddir setup, create gitignore/hgignore files
Meson does this too, and it is very useful. No one ever actually wants
to check their build directory into git, and if they did then they
should be taught the error of their ways. :p

Since build directories are completely arbitrarily named, it's not
practical or nice looking to check the names of build directories into
git as the project's gitignore file, but it's completely trivial for
muon to know when it is creating a build directory, and create these
files too.