restructure object memory layout
Rather than store all objects in a single array, use a different array
per object type.  This has the cost of some added complexity, as well as
an added indirection.  The benefits are greatly decreased memory usage,
especially for programs that use a disproportionate amount of array and
string objects.