reduce recursion in the interpreter
When debugging muon, the call stack can often grow to over 100 frames.
This is mostly due to the interpreter being a naive tree-walking
implementation combined with the AST being a very simple tree.

In particular, each expression is parsed to a node that has the left
child as the statement and the right child as the rest of the program.
This means that for every statement in a meson.build file, an additional
call frame gets pushed.  If your meson.build contains 1000 statements
(not uncommon), the backtrace from the final statement will be kind of
ridiculous.  Other than adding debugging friction, this also wastes
stack space and adds a little bit of latency.

Rather than make the entire interpreter recursion-free by using
trampolines or a bytecode, use a simple TCO for the node_block case
described above.  This cuts down the stack traces significantly with
minimal extra complexity.