fail pkgconf init without path in static build
When building muon statically, it gets the PKG_CONFIG_PATH of the system
on which it is built.  Unfortunately, PKG_CONFIG_PATH is different on
many different linux distributions.  This means that using static muon
on other distros has a high chance of failing to find dependencies, with
a non-obvious reason.

Improve the situation by refusing to intialize libpkgconf if there is no
PKG_CONFIG_PATH in the environment.