set pic: true by default for both_libraries
If b_staticpic was set to false, and pic was set to false or omitted,
the static library was compiled without -fpic.  Since both_libraries
reuses object files for the shared library, this resulted in linking
errors.  This solution sets the pic: kwarg to true for both_libraries if
it wasn't already set.