pkgconf: dont search filter_libdirs for libraries
filter_libdirs is pkgconf's idea of the search paths for system
libraries.  Originally, muon used this list to search for the library
before checking in directories specified with -L.  In reality,
filter_libdirs is what pkgconf uses to remove system libraries from the
output of --libs.  Since muon didn't invoke the filtering function, this
was redundant, all of the needed system paths should be specified with
-L (as if PKG_CONFIG_ALLOW_SYSTEM_LIBS=1 was set).

Additionally, filter_libdirs sometimes contained wrong paths that caused
incompatible libraries to be found.