clear temporary objects during backend output
After profiling, most of muon's memory usage appeared to be during
backend output, especially in the form of temporary strings (due to
building commandlines).  Since each target output function is pure, we
know that none of the objects created during target output will be used
for the next target.  Based on this assumption it was a simple matter to
implement a system whereby you mark the current level of the object pool
at the beginning of target output, and then reset it back to that level
at the end.

This strategy reduced muon's memory usage in one case by 1/3 (150mb ->
50mb), which puts it back under meson's memory usage.