optimize build.ninja depfiles
We don't need the extra DEPFILE / DEPFILE_UNQUOTED variable since we can
just construct that from within the build.ninja.

For meson's rationale for DEPFILE_UNQUOTED see commit
5ca37e7961ac2068382c36424c3f972ed0a9a105.  The rationale for DEPFILE,
however, I'm unable to discern.  It seems to have been the original
implementation, and I'm not sure if there are any benefits.