bootstrap: error out when using the bootstrap binary to build not-muon
The bootstrap binary is solely intended for use building muon itself,
and is missing significant functionality that is expected in the final
artifact.

Sometimes, people use bootstrap.sh and don't notice that it has crippled
functionality, so they forget to rebuild muon itself with it. Ensure
this is not possible, by asserting that the project name is "muon" if
running the bootstrap binary. This makes it:
- impossible to use bootstrap-muon for *anything* other than building
  muon
- obvious what to do to solve the problem