Python: query interpreter for more information.
This commit adds the necessary machinery to allow the
new python_installation object to query a Python interpreter
for information about its details.

Meson does this by passing a script to the interpreter dumps
its info to stdout as json, which Meson then parses.

I have copied the script that Meson uses and embedded it into muon.
This script will need to be kept up to date with Meson's version.

I have also added jsmn(https://github.com/zserge/jsmn), a small JSON
parser with zero dependencies into muon to allow the parsing of the
script's stdout data.

I started this work because I wanted to give muon the ability to
use python_installation.extension_module and some of the information
that is queried is used for that.