compute_int: Fix calculation of LONG_MAX
systemd computes the value of LONG_MAX according
to the C compiler and asserts if the value looks fishy.

muon currently produces a fishy value because it gets wrapped
to -1 due to the use of %d and not %ld in the string passed
to the compiler.

Fixing this allows muon to get closer to building systemd.