python: Expand type options for find_installation kwarg
Meson's documentation states that the `required` kwarg on
import('python').find_installation can be either a bool,
or, since version 0.48.0, a feature.

Currently it only accepts bool, which causes muon to raise a type
error when building systemd which passes a feature to this kwarg
at one point.

The commit fixes this, enabling both bools and features to be passed
to this `kwarg`. This allows muon to get closer to building systemd.