analyzer group source files by label
Originally the analyzer also grouped by the pointer to the source
structure.  I forgot why I originally did this but I was likely worried
about multiple sources having the same label and wanted an additional
way to differentiate them.  This seems unlikely to be helpful in
practice and is an issue now that functions use a temporary source
when they are called.