interpreter: Correct return value state handling
In cases where a script module function returns
void, while also calling an internal function which
does not return void, the return type of the internal
function ends up as the expected return type of the
outer void-returning function, yielding type failure.

Admittedly this is a very contrived example and only
likely to show up when prototyping new script modules
but it was nonetheless causing me difficulties in
exactly that situation so I have fixed it and
extended the script_module test to cover this case.