remove need for static linker rule to run in shell
The static linker rule looks like:

rm -f $out && ar crsD $out $in

It appears to be an ugly workaround for the fact that ar has no option
to truncate an archive.  This also doesn't work on windows because of
how ninja interprets the command.  Sidestep the issue by making one way
that should work on all platforms: add a flag to muon internal exe to
remove a file before running a command and then use that to invoke the
static linker:

muon internal exe -R $out -- ar csrD $out $in