refactor machine handling
machine info is now cached in a global struct per machine kind (build,
host).  This means code doesn't have to call machine_system() every time
it wants to know if the host machine is windows.  It also helps pave the
way for cross compilation.  Finally, a few stubs are inserted that can
further refine the host machine based on compiler interactions.  This
information is not yet put to use, however.