compilers.c: Fix detection of msvc
For some unknown and perhaps eternally unknowable
reason, some versions of msvc print the version
number to stderr instead of stdout. So in order to
correctly detect these versions of msvc we need to
check both stderr and stdout.

This commit add an extra check to do just that.
Meson seems to have already encountered this and
has an equivalent fix.