rename void to null
The name void was inherited from meson specifically to indicate
functions that don't return anything.

Technically in muon, all functions return something (just like in python
and many other scripting languages).  Not returning is equivalent to
returning null.  meson forbids assigning values to null understandably,
but this introduces difficulties when you want to e.g. tell if a kwarg
was set or not.  So muon added is_void to tell if a value is null.  This
got too confusing so I'm renaming everything to null.