implement copy on write for dicts
This allows a "light" dup to be made of a dict, which does not copy its
elements.  This dup will only copy elements when written to.  A light
dup is now made when you assign a dict to a var, rather than a full dup.
In addition a light dup is made during dict iteration to protect against
mutating a dict while it is being iterated over.

Fixes github #105.

Thanks again to @apprehensions for discovering this.