split array internal representation
Use two object types for arrays rather than just one.  This is similar
to how small dicts work.  obj_array remains the primary type, and a new
subordinate type is added obj_array_elem.  This simplifies the
bookeeping for arrays as well as potentially reduces memory usage since
before each array element had a len, head, and tail field whereas now
array elements contain only the value and a pointer to the next item.

Also, implement a cow strategy and use it in the same way as it is for
dicts.  This fixes array mutation during iteration.