tests/lang/run_command: use uppercase env vars
Managing envvars is annoying on windows.  run_cmd uses
SetEnvironmentVariable to build up the environment which converts all
variables to uppercase.  Apparently this is because environment
variables are case-insensitive on windows.  Python has a similar
behavior where all envvar accesses on windows are uppercased.

It should be possible to support lowercase envvars by manually
constructing the environment but I'm going to defer that work until it
breaks a project since it seems that technically that is "undefined
behavior".