communicate ninja to the test runner without split
The path to muon may contain spaces, so a simple split(' ') is
insufficient.  One approach would be to split on a different character
but we are already passing muon's path to the runner so just use a
special name for this case instead.

Thanks to @kzc for identifying this in github #127