add pkg-config-exec backend for pkg-config support
this adds a 'pkg-config-exec' backend which runs 'pkg-config'
binary (override via $PKG_CONFIG) at runtime and parses the
result rather than using libpkgconf.

libpkgconf option is removed and turned into a 'pkgconfig' combo
for selecting the choice of backend. default is set to 'auto'
which tries to use libpkgconf if found but falls back to 'exec'
otherwise.

Closes: https://github.com/muon-build/muon/issues/32