update pkgconf get_variable api
get_variable now passes in the list of defines avoiding global state.

additionally the libpkgconf implementation was broken since it never
reinitialized the pkgconf_client.  This was because older versions of
libpkgconf broke if this was attempted.  It appears to be fixed now so
the libpkgconf backend now inits a fresh client per pkgconf query.

This also requires a minimum libpkgconf version of 1.9.0.