unify regen deps handling and add exclude
If you have a subproject "foo" with a patch directory that patches in a
meson.build, like this:

subprojects/
- packagefiles/foo/meson.build
- foo/meson.build

Now muon will only mark packagefiles/foo/meson.build as a regen
dependency.  This is intended because every time muon executes a project
with patch files, it applies the patch directory.