fix installation failing for running executables
we can't overwrite a running binary file since we'll get
ETXTBSY. write to a tmpfile and do a rename() instead which
doesn't suffer from such issue.

this also coincidentally improves the robustness a bit since
previously a read/write failure would overwrite an existing file
and leave it broken. now we don't touch existing files until
writing is complete and so there's no chance of corrupting an
existing file.

Fixes: https://github.com/muon-build/muon/issues/138