mark some dependencies as valid for either machine
Dependencies track which machine they are for so that build/host
dependencies aren't accidentally mixed.  Some deps that only consist of
compile args, etc. are valid for both machines so they get an either
type.  Prior to this they would default to host and build machine
targets would be unable to use them.