use shell_split on certain options
These options have gotten a new option type: shell_array.  This is
almost identical to an array except that in the case the option is set
from an environment variable, or if the option is set on the command
line without using array syntax, it gets split according to shell
quoting rules.

Now these options are handled more similarly to meson, but we are more
explicit about it in that the option type must be "shell_array" to get
this handling and you can see which options have this type using `muon
options -a`.