reject options with source == option->source
The previous policy was to reject options with source < option->source.
This allowed options with the same source to override previous options.
For example -Da=1 -Da=2 would result in option 'a' being equal to 2.

This is wrong for default_options however, where you don't want a
subproject's default options to ovveride the parent project's.